+++ src/main.cpp
@@ -183,7 +183,7 @@ void runAppLoop() {
 void loop() {
 	logger::loop();
 	jsonRpc::loop();
-	if (!jsonRpc::hasPinConflict() || !jsonRpc::inUse()) {
+	if (!jsonRpc::inUse()) {
 		runAppLoop();
 	}
 }